Update IAVL tree in state and fix various issues
Improve logging and memomize transactions hashes
