+++ Gopkg.toml
@@ -1,13 +1,7 @@
-required = ["github.com/tendermint/iavl"]
-
 [prune]
   go-tests = true
   unused-packages = true
   non-go = true
-
-[[constraint]]
-  name = "github.com/tendermint/iavl"
-  version = "~0.2.0"
 
 [[constraint]]
   name = "github.com/BurntSushi/toml"
@@ -72,6 +66,10 @@ required = ["github.com/tendermint/iavl"]
 [[constraint]]
   name = "github.com/tendermint/go-wire"
   version = "~0.7.2"
+
+[[constraint]]
+  name = "github.com/tendermint/iavl"
+  version = "~0.5.0"
 
 [[constraint]]
   name = "github.com/tendermint/merkleeyes"
