+++ execution/namereg_cache_test.go
@@ -18,6 +18,7 @@ import (
 	"testing"
 )
 
+// TODO: tests
 func TestNewNameRegCache(t *testing.T) {
 }
