+++ execution/state.go
@@ -16,7 +16,6 @@ package execution
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
 	"sync"
@@ -25,71 +24,52 @@ import (
 	acm "github.com/hyperledger/burrow/account"
 	"github.com/hyperledger/burrow/binary"
 	"github.com/hyperledger/burrow/genesis"
-	"github.com/hyperledger/burrow/logging"
 	logging_types "github.com/hyperledger/burrow/logging/types"
 	"github.com/hyperledger/burrow/permission"
 	ptypes "github.com/hyperledger/burrow/permission"
 	"github.com/hyperledger/burrow/txs"
 	"github.com/tendermint/go-wire"
-	"github.com/tendermint/merkleeyes/iavl"
+	"github.com/tendermint/iavl"
 	dbm "github.com/tendermint/tmlibs/db"
-	"github.com/tendermint/tmlibs/merkle"
+)
+
+const (
+	defaultCacheCapacity = 1024
+
+	// Version by state hash
+	versionPrefix = "v/"
+
+	// Prefix of keys in state tree
+	accountsPrefix = "a/"
+	storagePrefix  = "s/"
+	nameRegPrefix  = "n/"
 )
 
 var (
-	stateKey                     = []byte("ExecutionState")
-	defaultAccountsCacheCapacity = 1000 // TODO adjust
+	accountsStart, accountsEnd []byte = prefixKeyRange(accountsPrefix)
+	storageStart, storageEnd   []byte = prefixKeyRange(storagePrefix)
+	nameRegStart, nameRegEnd   []byte = prefixKeyRange(nameRegPrefix)
 )
-
-// TODO
-const GasLimit = uint64(1000000)
-
-//-----------------------------------------------------------------------------
-
-// NOTE: not goroutine-safe.
-type State struct {
-	sync.RWMutex
-	db            dbm.DB
-	accounts      merkle.Tree // Shouldn't be accessed directly.
-	nameReg       merkle.Tree // Shouldn't be accessed directly.
-	lastSavedHash []byte
-	logger        logging_types.InfoTraceLogger
-}
 
 // Implements account and blockchain state
 var _ acm.Updater = &State{}
 var _ acm.StateIterable = &State{}
 var _ acm.StateWriter = &State{}
 
-type PersistedState struct {
-	AccountsRootHash []byte
-	NameRegHash      []byte
+type State struct {
+	sync.RWMutex
+	db      dbm.DB
+	version uint64
+	// TODO:
+	tree   *iavl.VersionedTree
+	logger logging_types.InfoTraceLogger
 }
 
-func newState(db dbm.DB) *State {
+func NewState(db dbm.DB) *State {
 	return &State{
-		db:       db,
-		accounts: iavl.NewIAVLTree(defaultAccountsCacheCapacity, db),
-		nameReg:  iavl.NewIAVLTree(0, db),
+		db:   db,
+		tree: iavl.NewVersionedTree(defaultCacheCapacity, db),
 	}
-}
-
-func LoadOrMakeGenesisState(db dbm.DB, genesisDoc *genesis.GenesisDoc,
-	logger logging_types.InfoTraceLogger) (*State, error) {
-
-	logger = logging.WithScope(logger, "LoadOrMakeGenesisState")
-	logging.InfoMsg(logger, "Trying to load execution state from database",
-		"database_key", stateKey)
-	state, err := LoadState(db)
-	if err != nil {
-		return nil, fmt.Errorf("error loading genesis state from database: %v", err)
-	}
-	if state != nil {
-		return state, nil
-	}
-
-	logging.InfoMsg(logger, "No existing execution state found in database, making genesis state")
-	return MakeGenesisState(db, genesisDoc)
 }
 
 // Make genesis state from GenesisDoc and save to DB
@@ -98,7 +78,7 @@ func MakeGenesisState(db dbm.DB, genesisDoc *genesis.GenesisDoc) (*State, error)
 		return nil, fmt.Errorf("the genesis file has no validators")
 	}
 
-	state := newState(db)
+	state := NewState(db)
 
 	if genesisDoc.GenesisTime.IsZero() {
 		// NOTE: [ben] change GenesisTime to requirement on v0.17
@@ -117,11 +97,10 @@ func MakeGenesisState(db dbm.DB, genesisDoc *genesis.GenesisDoc) (*State, error)
 			Balance:     genAcc.Amount,
 			Permissions: perm,
 		}
-		encodedAcc, err := acc.Encode()
+		err := state.UpdateAccount(acc.Account())
 		if err != nil {
 			return nil, err
 		}
-		state.accounts.Set(acc.Address.Bytes(), encodedAcc)
 	}
 
 	// global permissions are saved as the 0 address
@@ -137,11 +116,10 @@ func MakeGenesisState(db dbm.DB, genesisDoc *genesis.GenesisDoc) (*State, error)
 		Balance:     1337,
 		Permissions: globalPerms,
 	}
-	encodedPermsAcc, err := permsAcc.Encode()
+	err := state.UpdateAccount(permsAcc.Account())
 	if err != nil {
 		return nil, err
 	}
-	state.accounts.Set(permsAcc.Address.Bytes(), encodedPermsAcc)
 
 	// IAVLTrees must be persisted before copy operations.
 	err = state.Save()
@@ -153,21 +131,21 @@ func MakeGenesisState(db dbm.DB, genesisDoc *genesis.GenesisDoc) (*State, error)
 }
 
 // Tries to load the execution state from DB, returns nil with no error if no state found
-func LoadState(db dbm.DB) (*State, error) {
-	state := newState(db)
-	buf := db.Get(stateKey)
-	if len(buf) == 0 {
-		return nil, nil
+func LoadState(db dbm.DB, hash []byte) (*State, error) {
+	versionBytes := db.Get(prefixedKey(versionPrefix, hash))
+	if versionBytes == nil {
+		return nil, fmt.Errorf("could not retrieve version corresponding to state hash '%X' in database", hash)
 	}
-	persistedState, err := Decode(buf)
+	state := NewState(db)
+	state.version = binary.GetUint64BE(versionBytes)
+	err := state.tree.Load()
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("could not load versioned state tree")
 	}
-	state.accounts.Load(persistedState.AccountsRootHash)
-	state.nameReg.Load(persistedState.NameRegHash)
-	err = state.Save()
-	if err != nil {
-		return nil, err
+
+	if state.tree.LatestVersion() != state.version {
+		return nil, fmt.Errorf("state tree version %v expected for state hash %X but latest state tree version "+
+			"loaded is %v", state.version, hash, state.tree.LatestVersion())
 	}
 	return state, nil
 }
@@ -175,74 +153,29 @@ func LoadState(db dbm.DB) (*State, error) {
 func (s *State) Save() error {
 	s.Lock()
 	defer s.Unlock()
-	s.accounts.Save()
-	s.nameReg.Save()
-	encodedState, err := s.Encode()
+	s.version++
+	hash, err := s.tree.SaveVersion(s.version)
 	if err != nil {
 		return err
 	}
-	s.db.SetSync(stateKey, encodedState)
-	s.lastSavedHash = s.hash()
+	versionBytes := make([]byte, 8)
+	binary.PutUint64BE(versionBytes, s.version)
+	s.db.SetSync(prefixedKey(versionPrefix, hash), versionBytes)
 	return nil
-}
-
-func (s *State) LastSavedHash() []byte {
-	return s.lastSavedHash
-}
-
-func (s *State) Encode() ([]byte, error) {
-	persistedState := &PersistedState{
-		AccountsRootHash: s.accounts.Hash(),
-		NameRegHash:      s.nameReg.Hash(),
-	}
-	encodedState, err := json.Marshal(persistedState)
-	if err != nil {
-		return nil, err
-	}
-	return encodedState, nil
-}
-
-func Decode(encodedState []byte) (*PersistedState, error) {
-	persistedState := new(PersistedState)
-	err := json.Unmarshal(encodedState, persistedState)
-	if err != nil {
-		return nil, err
-	}
-	return persistedState, nil
-}
-
-// CONTRACT:
-// Copy() is a cheap way to take a snapshot,
-// as if State were copied by value.
-// TODO [Silas]: Kill this with fire it is totally broken - there is no safe way to copy IAVLTree while sharing database
-func (s *State) copy() *State {
-	return &State{
-		db:       s.db,
-		accounts: s.accounts.Copy(),
-		nameReg:  s.nameReg.Copy(),
-	}
 }
 
 // Computes the state hash, also computed on save where it is returned
 func (s *State) Hash() []byte {
 	s.RLock()
 	defer s.RUnlock()
-	return s.hash()
-}
-
-// As Hash without lock
-func (s *State) hash() []byte {
-	return merkle.SimpleHashFromMap(map[string]interface{}{
-		"Accounts":     s.accounts,
-		"NameRegistry": s.nameReg,
-	})
+	return s.tree.Hash()
 }
 
 // Returns nil if account does not exist with given address.
 func (s *State) GetAccount(address acm.Address) (acm.Account, error) {
 	s.RLock()
 	defer s.RUnlock()
-	_, accBytes, _ := s.accounts.Get(address.Bytes())
+	_, accBytes := s.tree.Get(prefixedKey(accountsPrefix, address.Bytes()))
 	if accBytes == nil {
 		return nil, nil
 	}
@@ -250,39 +183,31 @@ func (s *State) GetAccount(address acm.Address) (acm.Account, error) {
 }
 
 func (s *State) UpdateAccount(account acm.Account) error {
-	// TODO: interop with StateCache by performing an update on the StorageRoot here if storage is dirty
-	// we need `dirtyStorage map[acm.Address]bool`
-	//if dirtyStorage[account] == true {
-	//	s.accountStorage(account.Address())
-	//	 := acm.AsMutableAccount(account).SetStorageRoot()
-	//}
 	s.Lock()
 	defer s.Unlock()
-	encodedAccount, err := account.Encode()
+	// TODO: find a way to implement something equivalent to this so we can set the account StorageRoot
+	//storageRoot := s.tree.SubTreeHash(prefixedKey(storagePrefix, account.Address().Bytes()))
+	// Alternatively just abandon and
+	accountWithStorageRoot := acm.AsMutableAccount(account).SetStorageRoot(nil)
+	encodedAccount, err := accountWithStorageRoot.Encode()
 	if err != nil {
 		return err
 	}
-	s.accounts.Set(account.Address().Bytes(), encodedAccount)
+	s.tree.Set(prefixedKey(accountsPrefix, account.Address().Bytes()), encodedAccount)
 	return nil
 }
 
 func (s *State) RemoveAccount(address acm.Address) error {
 	s.Lock()
 	defer s.Unlock()
-	s.accounts.Remove(address.Bytes())
+	s.tree.Remove(prefixedKey(accountsPrefix, address.Bytes()))
 	return nil
-}
-
-// This does not give a true independent copy since the underlying database is shared and any save calls all copies
-// to become invalid and using them may cause panics
-func (s *State) GetAccounts() merkle.Tree {
-	return s.accounts.Copy()
 }
 
 func (s *State) IterateAccounts(consumer func(acm.Account) (stop bool)) (stopped bool, err error) {
 	s.RLock()
 	defer s.RUnlock()
-	stopped = s.accounts.Iterate(func(key, value []byte) bool {
+	stopped = s.tree.IterateRange(accountsStart, accountsEnd, true, func(key, value []byte) bool {
 		var account acm.Account
 		account, err = acm.Decode(value)
 		if err != nil {
@@ -293,62 +218,26 @@ func (s *State) IterateAccounts(consumer func(acm.Account) (stop bool)) (stopped
 	return
 }
 
-//-------------------------------------
-// State.storage
-
-func (s *State) accountStorage(address acm.Address) (merkle.Tree, error) {
-	account, err := s.GetAccount(address)
-	if err != nil {
-		return nil, err
-	}
-	if account == nil {
-		return nil, fmt.Errorf("could not find account %s to access its storage", address)
-	}
-	return s.LoadStorage(account.StorageRoot()), nil
-}
-
-func (s *State) LoadStorage(hash []byte) merkle.Tree {
-	s.RLock()
-	defer s.RUnlock()
-	storage := iavl.NewIAVLTree(1024, s.db)
-	storage.Load(hash)
-	return storage
-}
-
 func (s *State) GetStorage(address acm.Address, key binary.Word256) (binary.Word256, error) {
 	s.RLock()
 	defer s.RUnlock()
-	storageTree, err := s.accountStorage(address)
-	if err != nil {
-		return binary.Zero256, err
-	}
-	_, value, _ := storageTree.Get(key.Bytes())
+	_, value := s.tree.Get(prefixedKey(storagePrefix, address.Bytes(), key.Bytes()))
 	return binary.LeftPadWord256(value), nil
 }
 
 func (s *State) SetStorage(address acm.Address, key, value binary.Word256) error {
-	// TODO: not sure this actually works - loading at old hash
 	s.Lock()
 	defer s.Unlock()
-	storageTree, err := s.accountStorage(address)
-	if err != nil {
-		return err
-	}
-	if storageTree != nil {
-		storageTree.Set(key.Bytes(), value.Bytes())
-	}
+	s.tree.Set(prefixedKey(storagePrefix, address.Bytes(), key.Bytes()), value.Bytes())
 	return nil
 }
 
 func (s *State) IterateStorage(address acm.Address,
 	consumer func(key, value binary.Word256) (stop bool)) (stopped bool, err error) {
+	s.RLock()
+	defer s.RUnlock()
 
-	var storageTree merkle.Tree
-	storageTree, err = s.accountStorage(address)
-	if err != nil {
-		return
-	}
-	stopped = storageTree.Iterate(func(key []byte, value []byte) (stop bool) {
+	stopped = s.tree.IterateRange(storageStart, storageEnd, true, func(key []byte, value []byte) (stop bool) {
 		// Note: no left padding should occur unless there is a bug and non-words have been writte to this storage tree
 		if len(key) != binary.Word256Length {
 			err = fmt.Errorf("key '%X' stored for account %s is not a %v-byte word",
@@ -372,7 +261,7 @@ func (s *State) IterateStorage(address acm.Address,
 var _ NameRegIterable = &State{}
 
 func (s *State) GetNameRegEntry(name string) (*NameRegEntry, error) {
-	_, valueBytes, _ := s.nameReg.Get([]byte(name))
+	_, valueBytes := s.tree.Get(prefixedKey(nameRegPrefix, []byte(name)))
 	if valueBytes == nil {
 		return nil, nil
 	}
@@ -381,16 +270,9 @@ func (s *State) GetNameRegEntry(name string) (*NameRegEntry, error) {
 }
 
 func (s *State) IterateNameRegEntries(consumer func(*NameRegEntry) (stop bool)) (stopped bool, err error) {
-	return s.nameReg.Iterate(func(key []byte, value []byte) (stop bool) {
+	return s.tree.IterateRange(nameRegStart, nameRegEnd, true, func(key []byte, value []byte) (stop bool) {
 		return consumer(DecodeNameRegEntry(value))
 	}), nil
-}
-
-func DecodeNameRegEntry(entryBytes []byte) *NameRegEntry {
-	var n int
-	var err error
-	value := NameRegDecode(bytes.NewBuffer(entryBytes), &n, &err)
-	return value.(*NameRegEntry)
 }
 
 func (s *State) UpdateNameRegEntry(entry *NameRegEntry) error {
@@ -401,13 +283,30 @@ func (s *State) UpdateNameRegEntry(entry *NameRegEntry) error {
 	if err != nil {
 		return err
 	}
-	s.nameReg.Set([]byte(entry.Name), w.Bytes())
+	s.tree.Set(prefixedKey(nameRegPrefix, []byte(entry.Name)), w.Bytes())
 	return nil
 }
 
 func (s *State) RemoveNameRegEntry(name string) error {
-	s.nameReg.Remove([]byte(name))
+	s.tree.Remove(prefixedKey(nameRegPrefix, []byte(name)))
 	return nil
+}
+
+// Creates a copy of the database to the supplied db
+func (s *State) Copy(db dbm.DB) *State {
+	state := NewState(db)
+	s.tree.Iterate(func(key []byte, value []byte) bool {
+		state.tree.Set(key, value)
+		return false
+	})
+	return state
+}
+
+func DecodeNameRegEntry(entryBytes []byte) *NameRegEntry {
+	var n int
+	var err error
+	value := NameRegDecode(bytes.NewBuffer(entryBytes), &n, &err)
+	return value.(*NameRegEntry)
 }
 
 func NameRegEncode(o interface{}, w io.Writer, n *int, err *error) {
@@ -416,4 +315,28 @@ func NameRegEncode(o interface{}, w io.Writer, n *int, err *error) {
 
 func NameRegDecode(r io.Reader, n *int, err *error) interface{} {
 	return wire.ReadBinary(&NameRegEntry{}, r, txs.MaxDataLength, n, err)
+}
+
+func prefixedKey(prefix string, suffices ...[]byte) []byte {
+	key := []byte(prefix)
+	for _, suffix := range suffices {
+		key = append(key, suffix...)
+	}
+	return key
+}
+
+// Returns the start key equal to the bytes of prefix and the end key which lexicographically above any key beginning
+// with prefix
+func prefixKeyRange(prefix string) (start, end []byte) {
+	start = []byte(prefix)
+	for i := len(start) - 1; i >= 0; i-- {
+		c := start[i]
+		if c < 0xff {
+			end = make([]byte, i+1)
+			copy(end, start)
+			end[i]++
+			return
+		}
+	}
+	return
 }
