+++ logging/loggers/info_trace_logger.go
@@ -97,6 +97,7 @@ func (l *infoTraceLogger) Log(keyvals ...interface{}) error {
 // Wrap the output loggers with a a set of standard transforms, a non-blocking
 // ChannelLogger and an outer context
 func wrapOutputLogger(outputLogger kitlog.Logger) (kitlog.Logger, channels.Channel) {
-	return NonBlockingLogger(VectorValuedLogger(SortLogger(BurrowFormatLogger(outputLogger),
-		structure.ChannelKey, structure.MessageKey, structure.TimeKey, structure.ComponentKey)))
+	//return outputLogger, channels.NewDeadChannel()
+	return VectorValuedLogger(SortLogger(BurrowFormatLogger(outputLogger),
+		structure.ChannelKey, structure.MessageKey, structure.TimeKey, structure.ComponentKey)), channels.NewDeadChannel()
 }
