+++ txs/tx_test.go
@@ -237,34 +237,13 @@ func testTxMarshalJSON(t *testing.T, tx Tx) {
 	assert.Equal(t, string(bs), string(bsOut))
 }
 
-/*
-func TestDupeoutTxSignable(t *testing.T) {
-	privAcc := acm.GeneratePrivateAccount()
-	partSetHeader := types.PartSetHeader{Total: 10, Hash: makeAddress("partsethash")}
-	voteA := &types.Vote{
-		Height:           10,
-		Round:            2,
-		Type:             types.VoteTypePrevote,
-		BlockHash:        makeAddress("myblockhash"),
-		BlockPartsHeader: partSetHeader,
+func TestTxHashMemoizer(t *testing.T) {
+	tx := &CallTx{
+		Input: &TxInput{
+			Sequence: 4,
+		},
 	}
-	sig := privAcc acm.ChainSign(chainID, voteA)
-	voteA.Signature = sig.(crypto.SignatureEd25519)
-	voteB := voteA.Copy()
-	voteB.BlockHash = makeAddress("myotherblockhash")
-	sig = privAcc acm.ChainSign(chainID, voteB)
-	voteB.Signature = sig.(crypto.SignatureEd25519)
-
-	dupeoutTx := &DupeoutTx{
-		Address: makeAddress("address1"),
-		VoteA:   *voteA,
-		VoteB:   *voteB,
-	}
-	signBytes := acm.SignBytes(chainID, dupeoutTx)
-	signStr := string(signBytes)
-	expected := fmt.Sprintf(`{"chain_id":"%s","tx":[20,{"address":"%s","vote_a":%v,"vote_b":%v}]}`,
-		chainID, *voteA, *voteB)
-	if signStr != expected {
-		t.Errorf("Got unexpected sign string for DupeoutTx")
-	}
-}*/
+	hsh := tx.Hash("foo")
+	assert.Equal(t, hsh, tx.txHashMemoizer.txHashBytes)
+	assert.Equal(t, "foo", tx.txHashMemoizer.chainID)
+}
